Added xrun script and a basic argument parser
Added to xmenu the option "-c" to specify the command that must be used to
complete the user input.
Included the xrun script that implements an application laucher. It exports an
function that uses bash complete to give a list of completion options to xmenu.
This function is called by the xmenu through the option "-c".